backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "todo-backend"
version = "1.0.0"
description = "Todo application backend using FastAPI and SQLModel"
authors = [{name = "Developer", email = "dev@example.com"}]
dependencies = [
    "fastapi>=0.104.1",
    "sqlmodel>=0.0.16",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.5.0",
    "uvicorn>=0.24.0",
    "pytest>=7.4.3",
    "httpx>=0.25.2",
    "python-dotenv>=1.0.0",
    "passlib[bcrypt]>=1.0.0",
    "better-exceptions>=0.3.3",
    "openai-agents>=0.8.0",
    "mcp>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.6"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]
